Update myScript.js
Added the legend section to the Treemap and passes all the 12 test from the FCC suite. The project is now complete. For the moment the "Video Game Sales" - Dataset is the default one. I'll probably come back in the future to add the possibility to switch between the Video, Movie and Kickstarter datasets as an option for this project. Project submitted for review.